+++ src/main/java/br/com/grupomult/repositories/CarroRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.grupomult.entities.Carro;
 
 @Repository
-public interface CarroRepository extends CrudRepository<Carro, Integer> {
+public interface CarroRepository extends CrudRepository<Carro, Long> {
 
 }
